Revert "Merge pull request #7 from vimc/i1744"
This reverts commit ca09d317b33bae0bd32cb8fee0dee2d2cd572603, reversing
changes made to f88a044d9b8600e458138d9ad809425d93817928.

While this feature branched passed on teamcity, it is failing after merge
for reasons that are not clear.  Further poking about is not helping (see
the failing i1744_repoen branch currently failing on TC).  James is
going to investigate something that does not depend on rJava